pyproject.toml
[project]
name = "o2view"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Frederik Krämer", email = "frdrkkraemer@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "dash>=2.18.2",
    "dash-ag-grid>=31.3.0",
    "dash-mantine-components>=0.15.3",
    "flask-caching>=2.3.1",
    "numpy>=2.2.3",
    "plotly>=6.0.0",
    "polars[excel]>=1.24.0",
    "pyjanitor>=0.31.0",
    "pywebview>=5.4",
    "scipy>=1.15.2",
    "setproctitle>=1.3.5",
]

[project.scripts]
o2view = "o2view:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
